tut/quickstart/requirements.txt
asttokens==2.2.1
backcall==0.2.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
executing==1.2.0
idna==3.4
ipython==8.7.0
jedi==0.18.2
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
numpy==1.23.5
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
requests==2.28.1
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
torch==1.13.1+rocm5.2
torchaudio==0.13.1+rocm5.2
torchvision==0.14.1+rocm5.2
traitlets==5.7.1
typing-extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
